backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
pydantic>=2.6
pydantic-settings>=2.2
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pdfplumber>=0.11
pytesseract>=0.3.10
rapidfuzz>=3.6
regex>=2023.12
pandas>=2.0
psutil>=5.9
pillow>=10.0
